make AS create rooms instead of puppet client and bump minor version to 1.15.0
this change effectively makes the AS have the most control over the
room. although we set the puppet user's power level to 100, he won't
have the ability to, for example, remove the alias to the room.
this shift of power from the user to the AS is made rather ostensibly.

the main purpose of this change is to allow configuring the registration
file of the bridges to designate control over a room namespace and yet
still be able to create a room in that namespace. in other words,
installing the bridge would fail if the room alias namespace was
registered (e.g. #facebook_*) and the puppet client tried to create the
protocol room.

with this change, the protocol room is now created by the AS, which is
allowed to create rooms in that namespace

fixes #17
unblocks https://github.com/matrix-hacks/matrix-puppet-facebook/pull/24