using botIntent to create rooms
upgraded matrix-* dependencies

attempted to fix #18 but it appears the createAsClient
option is not actually honored, so the AS bot joins the room.

because it's better for the AS bot to manage the rooms, I am leaving
this as-is rather than going back to the design in which the puppet
client is the one creating the rooms.

again, the reason we need the AS bot to manage the rooms is so that the
AS can safely register #service_* room alias. if the puppet client tried
to create rooms with this setup, synapse will not allow it.